refactor(ui): colocate SSE event validation in api/validation (#43) Move OpenAPI SSE guards from utils into api/validation.ts with a note that they return null and warn instead of throwing ContractError for streams. Remove sseEventValidation.ts; keep parseSseFrame malformed-JSON console.warn and tests alongside api validation tests.
